# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "personal projects",
        "title": "deja server",
        "termsOfService": "https://github.com/nanzm",
        "contact": {
            "name": "nan",
            "url": "https://nancode.cn",
            "email": "msg@nancode.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/comment/new": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "添加评论",
                "parameters": [
                    {
                        "description": "评论",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/comments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一页数量",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/movie/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "电影详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "电影id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/movie/log": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "添加观看记录",
                "parameters": [
                    {
                        "description": "电影id",
                        "name": "movie_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "进度 00:02:32",
                        "name": "progress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/movie/logs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "观看记录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一页数量",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/movies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "电影列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一页数量",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/upload": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一页数量",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/weapp/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "code登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wx.login 获取的 code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "object"
                },
                "status": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}